git-commit-setup-check-buffer: Autoload
This allows using "git commit" from the command-line to connect to
a running Emacs instance in which the server has been started and
in which only "magit-autoloads.el" (but not "git-commit.el" and its
dependencies) have been loaded already loaded.

  $ cd /path/to/magit
  $ emacs -Q ... --load lisp/magit-autoloads.el \
    --eval "(add-hook  'find-file-hook 'git-commit-setup-check-buffer)" \
    --funcall server-start &

  $ cd /path/to/repo
  $ GIT_EDITOR="emacsclient -t" git commit

"..." stands for whatever has to be done to put magit, magit-popup,
with-editor and dash on the `load-path'; in my case "-L lisp -L
../with-editor -L ../dash -L ../magit-popup".

Re #3415.