Refactor blame commands
Previously the required information was determined in one place and
then passed along to the place where it was needed.  More specifically
all values were determined within the commands' `interactive' form and
then used in `magit-blame--run'.  Additionally the code used to
collect the necessary information was interweaved with error detection.

Now the commands take no arguments at all and call `magit-blame--run'
with a single argument, TYPE, which essentially identifies them.

The task of visiting the appropriate blob in a buffer (when necessary)
and the task of adding blame information to that buffer are clearly
separated.  The commands now
* call `magit-blame--pre-blame-assert' to avoid trying to blame when
  that is not possible;
* call `magit-blame--pre-blame-setup' to visit the appropriate blob,
  if that is necessary, and to jump to the correct line; and
* call `magit-blame--run' to add the blame information to the buffer.

When the current buffer does not already contain blame information,
then `magit-blame--pre-blame-setup' does nothing.  If it needs to
visit a blob, then it delegates that to `magit-blame-visit-prev-file'.
When the "blaming direction" changes, then it clears existing
overlays.