magit-blame-type: Set in magit-blame--pre-blame-setup
Previously this was done in `magit-blame--run', which also worked
but is weird - the purpose of this function is to run "git blame"
according to the configuration and not to setup the configuration.

This allows removing `magit-blame--run's TYPE argument, which allows
adding it directly to `after-save-hook', which in turn makes the
wrapper `magit-blame-after-save-refresh' unnecessary.