magit-gitk-executable: fix default gitk filename, on Windows
We were accidentally trying to look for "gitk" in a directory called
"git.exe", instead of searching for "gitk" in the directory containing
the git executable.  With Emacs 27, this fails with an error (earlier
versions fail silently): Tramp tries to extract a file "gitk" from an
archive "git.exe", and signals error "package 'tramp-archive' not
supported".