magit-blame-goto-chunk-hook: Bugfix
When blaming is still in progress, then movement might be from a
blamed chunk to a yet-to-be blamed chunk, in which case we probably
should not remember `nil' as the previous chunk.  That's from a
technical perspective - from a usage perspective there can be
different opinions on what is better.