Fix remembering whether user confirmed editing published commits
We now use a variable to remember whether the user already confirmed
that s/he wants to edit published commits when initiating the rebase
so that s/he won't be asked again when there are staged changes when
continuing the rebase.  Previously we tried to use a file in the
control directory used by Git, but that directory does not exist at
the time we need to remember and if we create it ourselves, then Git
trips over it.

Fixes #3447.