magit-blame--format-time-string: Correct timezone handling
Because the time variable is an Unix time which is not related to time zone,
we don't have to update time according to time zone.

To deal with the different behavior of `format-time-string`

In emacs version > 25 :

If time-format contains the time zone symbol ( %z or %Z ),
use tz-in-second to format the time in committer time zone or author time zone.

If time-format doesn't contain the time zone symbol ( %z or %Z ),
assign tz-in-second to nil to format the time in local time zone.

In emacs version < 25 :

Just format the time string in local time zone