Avoid error if gitk not found on windows-nt
On windows-nt, if gitk is not in the path, then the 'which' produces
nothing to stdout (although it does produce an error message to
stderr).  The magit-git-string call then returns nil, which is not a
valid input to file-executable-p, causing an error when trying to
autoload anything defined in magit-extras.el.

Include 'exe' as a first conjunct to the 'and', short-circuiting the
file-executable-p test in this case.