magit-rebase-autosquash: Skip usual asserts
The asserts were already run when the fixup and squash commits were
created, so we don't have to ask again.  Furthermore we always rebase
all commits since the upstream, which means that (1) changes reachable
from upstream (including its tip) are never autosquashed into by this
command, and (2) that we might rebase over commits that actually are
unaffected in the end but that would be considered problematic.

Closes #3468.