magit-dired-jump: Don't append slash to directory
Since GNU Emacs 26.1, specifically commit [1: 3988ceaa1c] addressing
bug#24409, dired-jump passes a non-nil FILE-NAME argument to
expand-file-name.  The expansion removes the trailing "/."  which
magit-dired-jump previously appended to directory name arguments,
resulting in a jump to the parent of the desired directory.

In place of the "/." approach of Magit commit [2: 5665ba94], call
file-name-as-directory when magit-file-at-point returns a directory.
This is necessary because magit-file-at-point returns a submodule path
without a trailing slash, and we want to jump to the submodule
directory not the parent directory.

In addition, simplify magit-dired-jump by passing a FILE-NAME argument
of nil to dired-jump when no file is detected at point, as dired-jump
will default to default-directory in this case.

[1: 3988ceaa1c]: 2016-09-14 00:04:32 +0900
  dired-jump: Expand file-name before dired-goto-file call
  https://git.savannah.gnu.org/cgit/emacs.git/commit/?id=3988ceaa1cfa2022443906750cbb30127a1e161f

[2: 5665ba94]: 2016-10-06 19:07:30 +0200
  magit-dired-jump: visit directory, not parent directory
  https://github.com/magit/magit/commit/5665ba9406b40e4cab97866b220e404211a4bd90