magit-gitignore-popup: New popup
Bind "I" to this popup instead of to `magit-gitignore-locally'.
That command is now available at "I l".  `magit-gitignore' (the
global variant) is now available on "I g", but continues to be
available on "i" by default.

The purpose of this change is to make it possible to use "i"
for `magit-imerge-popup' from the `magit-imerge' package without
making `magit-gitignore' unavailable.

Discussed in https://github.com/magit/magit-imerge/issues/2.