Force standard source and destination prefixes in diffs
We do this for the same reason for which we used to force the use of
no prefix at all: in order to be able to correctly parse filenames in
diffs, we have to know whether a prefix is used and how long it is.

I probably chose "--no-prefix" in ebccd16914 to do so, because that
is only one argument.  Unfortunately Git in some cases ignores that
argument, namely when the "-L" argument is also used.

Forcing the use of the default prefixes instead should be safer
because it doesn't matter whether we get the desired output because
our request was honored or because the user's "diff.noprefix=true"
was ignored as well.  In both cases the default prefixes are used.

I am choosing "git -C diff.noprefix=false ..." over "git diff
--src-prefix=a/ --dst-prefix=b/" because that means we don't have to
make sure we use these arguments where ever "git diff" is called.