Allow specifying the face to use to highlight certain section
If the new `heading-highlight-face' slot is non-nil for a section,
then that is used to highlight its heading when it is selected.  This
is useful for sections whose headings have a background color when
they are not selected.  Without this the background would be covered
by that of the face that is always used to highlight the heading and
the body.

Something similar already exists for diff-related; this is a
simplified implementation for use by sections with fewer special
needs.