magit-branch-adjust-remote-upstream-alist: Fix snippet rendering
There are two issues with how this doc-string example is rendered in
the help buffer: (1) there is a single backslash but the regexp needs
two and (2) the grave accent may be displayed as ‘ depending on the
value of text-quoting-style.

Do all the (very ugly) escaping that's necessary to render the example
reliably in the help buffer.