magit-describe-sections-briefly: Cosmetics
We would like to use `and-let', but that isn't available in all
supported Emacs releases.  Among the choices we have, `when-let'
is better suited than `if-let' because there is no ELSE part.

This also addresses #3558 in passing (for a while the development
version did not support `if-let' an abbreviated single-variable
SPEC).