Only jump to commits when first creating the status buffer
Do so using the new `magit-create-buffer-hook'.

Doing this was the intention in [1: 1ec3d7110] but that ended up
jumping to the commits whenever `magit-status' was used.

Fixes #3561.

Note that it would have been more convenient for this feature if
the previous commit had delayed running the new hook until after
`magit-refresh-buffer' was run, but while that would have made
sense here, that would have made that hook less useful generally
speaking.

1: 1ec3d71103dcbc24eac1a00863a2923fea4b9ea3
   When showing a status buffer for the first time, jump to some log