magit-log-heading-re: Make the date field optional
'git rebase -i --root' corrupts the root commit's author field.  As a
result, the %at field for the root commit is empty, and we end up with
a log where the root commit remains unwashed, such as

    8c56a87 a
            c224010^@^@Kyle Meyer^@^@b

(This corruption has been fixed in what will be Git v2.19.0 by commits
ca3e1826a0, 0f16c09aae, and 67f16e3d3f.)

Update magit-log-heading-re to match an empty date field so that we
can detect and parse the line for a corrupted root commit.

Closes #3574.