magit-section-when: Stop using it as originally intended
Only three functions used `magit-section-when' as originally intended,
indicating that it probably isn't as useful as imagined.  Another
function uses it a strange way that isn't used anywhere else, even
though `magit-section-when's doc-string hints at such a use-case.

Stop using `magit-section-when' like that, so that it can be
deprecated and be replaced with a function that better suits the
majority of callers.  The remaining callers use `magit-section-when'
in a way that does not justify the "when" in its name.