magit-section-lineage: Return list instead of vector
The list is the "natural form".  The vector is only needed when
showing the lineage to the user because that makes it easier to make
the connection to condition vectors as used by `magit-section-match',
which has to use vectors for lineages because it already uses lists
to represent lists of multiple conditions.