magit-section-match: Support classes in condition
Previously the condition had to consist of type symbols, now it
can also contain class symbols.  This makes it more convenient
to use section types that derive from another.  Previously that
was cumbersome because each caller that wanted to match a class
and its subclasses had to match each subclass explicitly.