magit--record-separated-gitdir: Resolve symlinked gitdirs
git-annex converts submodule gitdir files to symlinks.  As a result,
magit-toplevel doesn't discover the correct working tree when it's
called from .git/modules/NAME.  The main user-visible effect is that
magit-diff-while-committing shows empty diffs.

We could address this by adding a kludge to magit-toplevel, but that
would mean risking regressions in a tricky and core part of the magit
codebase to work around the peculiar behavior of a third-party
program.

However, this is similar to the problem that was solved in
439ed762 (Record the working tree for separated git directories,
2017-02-05).  We can piggyback on that solution by teaching
magit--record-separated-gitdir to resolve symlinked gitdirs. In the
worst case scenario, we incorrectly identify the gitdir.  Then
magit-toplevel, when called from the correct gitdir, simply wouldn't
find a corresponding entry in magit--separated-gitdirs, and we'd end
up where we are without this change.

Fixes #3599.