magit-bisect-start: Error out when there are uncommitted changes
Strictly speaking it is possible to bisect with uncommitted changes,
provided these changes don't touch the same files as the commits that
are being bisected.  Otherwise checkout fails and an error is shown
in the status buffer.  However if the very first checkout fails due
to this, then the failure is silent and to avoid that we disallow
bisecting with uncommitted changes altogether.

Fixes #3604.