magit-{stage,unstage}-hook: New hooks
Running these hooks inside `magit-refresh' has the benefit, that
the refresh cache is used, possibly making expensive calculations
unnecessary.  Using variables to control the commands for which
the hooks should be called has the benefit that it is easy to add
new commands (including third-party commands) without having to
complicate their implementations.

See #3496.