Require magit-submodule when loading magit
The keymap for the `module' section type is defined in
`magit-submodule' and used by a section inserter in `magit'.  Due to
how `magit-insert-section' uses section keymaps we cannot rely on the
variable being autoloaded - calling `boundp' does not trigger the
library that defines the autoloaded variable to be loaded.

Fixes #3618.