magit-repository-directories: Add some default directories
If these directories do not actually exist or don't contain any
repositories, then that isn't much of a problem but it helps users of
`borg'.  Users who want to use that features that use the value of
this option have to customize this anyway (except now `borg' users),
so the default being non-nil shouldn't upset anyone.