magit-reverse-files: Guard against empty section list
magit-reverse-files removes binary files from `sections', so it may
end up nil.  At the time this code was introduced, passing nil to
magit-reverse-apply wasn't an issue, but it leads to a type error in
magit-apply-patch since ddce6ce4 (sections: Use oref instead of
dedicated slot accessors, 2018-01-17).

Note: magit-discard-files--discard performs a similar removal of
binary files, but it doesn't lead to a type error because it already
guards against the nil case.