magit-with-editor-envvar: New variable
Setting this to `GIT_SEQUENCE_EDITOR' means that Emacs will be used to
edit a file on behalf of `git rebase -i' but not when creating commits.
In that case `GIT_EDITOR' (if set, else `EDITOR') will be used to edit
commit messages and similar files.  Setting one of these env variables
is the responsibility of the user; by setting this variable, they have
opted out of having Magit do it for them.

Closes #3629.