magit-staged-binary-files: Generalize and rename
This function exists so that we can detect when a user is attempting
to perform an invalid operation on a binary file and give an
informative error message.  To make the reverse checks more
comprehensive, we'll need to check diffs other than the one between
HEAD and the index.  In preparation for this, adjust
magit-staged-binary-files to take an arbitrary list of of git-diff
arguments.