git-commit-post-finish-hook: Don't run for unsafe commands
When a command like "git rebase --continue" used `EDITOR' to have the
user provide a commit message and created a commit with the message it
received that way, then it goes on doing other things that may need to
lock the index.

We therefore cannot run `git-commit-post-finish-hook' after the commit
was created because otherwise there would be a race condition between
the subprocesses of "git rebase" and whatever git commands the hook
calls.  Blacklist commands known to keep going after committing.

This was previously done in `magit-wip-maybe-add-commit-hook', but we
need to prevent all of `git-commit-post-finish-hook' to be inhibited.