magit-gitignore: Clarify action of magit-gitignore-globally
Git already uses the terms "local" and "global" for matters pertaining
to configuration in a consistent way: "local" to mean the repository
config file, and "global" to mean the global (per-user)
configuration.  magit-gitignore used the term "global" to mean
"versioned" in a context where "global" could have well been
interpreted as per-user.

Avoid confusion by removing the ambiguous "global" term in the popup
action description, and explicitly clarifying which file will be
edited in the functions' documentation.