Revert "make: Tie git-commit's version to magit's version"
This reverts commit 2dbccdaadb6c5a69ff79165b45d635effcb2f281.

It's better not to do this unless absolutely necessary. After the new
`magit' release has made it onto Melpa it can take hours until the new
`git-commit' release is available too.  During that time `magit' would
be uninstallable.

Usually it is not necessary to bump the `git-commit' dependency
because nothing in "git-commit.el" changed.  In those cases we don't
want `magit' to potentially become uninstallable for a while.  (Well,
we never do, but here we can avoid without losing anything.)

Added a note to the release playbook instead.