magit-repository-directories: Use nil as default again
Because many users put "~/.emacs.d/" under version-control but don't
customize `magit-repository-directory', that directory should not be
a member of `magit-repository-directories'.  Otherwise `magit-status'
would not fall through to either offering to initialize the current
directory as a repository (`magit-init' would have to be used) or to
read a repository using regular file-name completion (except with two
prefix arguments), and users have grown accustomed to these fallback
behaviors.

Also update the documentation about `magit-repository-directories'
and `magit-status'.  The documentation about the latter was rather
ambiguous until now.  We trade that for making it a bit scary instead.

This reverts commit f77d1158b01923ca11d065f12aa0e69d6ce2415b, and
then also improves the documentation.

This fixes #3673.