Rename face git-commit-note to git-commit-keyword
In the rest of Magit we call [those] things "keywords".  Do so here
too.  Additionally we should avoid the term "note" because it means
something else in the context of editing text using `git-commit'.
A note is an additional text that is associated to a commit (similar
to a commit message, but unlike that not actually part of the commit
object) so the same word should not be used for a certain kind of
substring of a commit message.