magit-rev-verify: Use magit-git-string-p
Previously this used `magit-rev-parse-save', but that failed to
return nil if REV isn't a commit but a range A..B, because it used
`magit-git-string', which returns the first line of the output even
if the exit code is non-zero.

In that case the output would be something like:

  cc998e85c00dff5b2429155b1b763b0f6e38ef97    # stdout
  ^d2c84d9a56e3cf12127ca22f72273febd92538b7   # stdout
  fatal: Needed a single revision             # stderr

I would expect git to fail without any output in this case,
but that's not what happens.

Use the new `magit-git-string-p' instead; that returns nil
if the exit code is non-zero, even if there is output.