Use a macro when delaying hidden section bodies
Add a new macro `magit-section-on-expansion' and use it where ever
the creation of the section body is already being delayed when the
it would be hidden.

That makes it a bit easier to delay the creation of such bodies,
and it also makes it possible to add code that has to be run when
ever any such section is expanded.