magit-diff-inside-hunk-body-p: Handle type change hunks
Magit treats type changes as a sort of hunk.  Such hunks do not have
a body and we have to take this into account here because otherwise
we get a type error when trying to use its content slot as a number.
That in turn would prevent some other code from running resulting in
section movement commands getting stuck on this section after one of
`magit-previous-line' or `magit-next-line' was used inside the type
change hunk.

Closes #3698.