git-rebase.el: split git-rebase-exec
git-rebase-exec is able to insert an exec command in the next line of
the todo list, detecting if there already is such a command on the
current line and replacing the current command, prompting the user for
the said command, or dropping the command if the user typed nothing.

These features are also interesting for the label and the reset
command -- for instance, it is useless to put a reset command after
another one, or to put two labels at the same place.

This moves the main function of git-rebase-exec to a new function,
git-rebase-insert-at-point-with-args.  This function can be passed
with a regex to recognize the command to override, the command to be
inserted, the function to read the user’s argument, if the command is
to be inserted to the next line (for reset or exec), and the universal
argument.

Signed-off-by: Alban Gruin <alban@pa1ch.fr>