magit-rev-name: Use --exclude argument if available
We already used "--exclude" before but had to stop doing so because
that argument wasn't added until Git v2.13.  See [1: 7d61fe4e5] for
the story until now.

The benefit of this approach is that this function does not *always*
fail to return something when using an old Git version.  The drawback
is that when using a new Git version, then it *often* does not return
something even though it theoretically could return what we wanted it
to return.

Going forward use two implementations.  The >=2.13 implementation does
what we want to happen and the <2.13 implementation makes a best effort.

1: 7d61fe4e592cbc3483eda5a5f03d9017031b11b6
   magit-pull-from-upstream: Use magit-git-pull