[key] magit-patch-create: Bind mail arguments to "C-m ..."
We cannot bind "-C" to "--cc" because the former is the short argument
for "--find-copies".  We do not want to keep using "=c" because we are
trying to move away from that prefix across all transients.  We also
do not want to use "-c" because we try to use "-<uppercase>" when
moving away from "=<letter>".

Right now "--cc" is the only mail argument that conflicts with a diff
argument, as described above.  But that might changes and we want a
future proof solution.

So we use "C-m ..." for all mail arguments.  For consistency we do so
even for "--reroll-count" and "--output-directory", which actually
have short arguments, "-v" and "-o".