magit-insert-head-branch-header: Error on failed magit-rev-format
Fixes #3444 and #3276 (kinda)

If a repo is corrupt, magit-rev-format will return nil, and
string-match will raise a type error. This is confusing, making it
sound like something went wrong with Magit, when the true issue is
more dire. The error message in this commit is vague, but it at least
hints that something has gone wrong at the Git level. I don't think
it's possible to get any more specific while remaining unobtrusive.