git-rebase: Store line information in intermediate structure
The rebase actions all gather their information through the matches of
the git-rebase-line regexp.  This works fairly well right now; the one
hiccup is git-rebase-noop, which adds a trailing blank after the lines
it inserts so that it matches git-rebase-line.

But the possible rebase lines are going to get more complicated when
we add support for 'git rebase --rebase-merges' and for the break
keyword.  Instead of complicating the regexp and making each caller
deal with these details, introduce a new function
git-rebase-current-line that parses the line into a common structure
that callers can use.

Represent the line information as a class.  This is admittedly an
overkill because we don't need any of the object-oriented features,
but it makes access more convenient.