magit-rebase-insert-merge-sequence: Support --rebase-merges
Now that we can pass --rebase-merges for interactive rebases, we need
to be able to display several new action types: merge, reset, and
label.  Treat most of these actions in the same way as we treat exec
(i.e. don't insert a proper section) because there isn't an obvious
value to store/display.  The exception is merge actions that include
"-C/c <message hash>"; in that case, store the message hash as the
section value.

Use git-rebase.el's git-rebase-current-line so that we don't have to
duplicate all the action parsing logic here.

We may want to explore other ways of displaying the --rebase-merges
sequence because the single line display isn't ideal for showing the
branch structure, but for simple cases the current representation
gives a good enough idea and the user can always resort to other
methods (e.g., 'log REBASE_HEAD..ORIG_HEAD') to get a better view.