git-rebase-description: Correct doc-string
This face isn't applied to commented exec lines.  It seems sensible to
treat commented commit actions different than exec (or noop), so
adjust the documentation rather than the font-lock regexp.

Use the phrase "commit" action to prepare for --rebase-merge
actions (reset, label, and merge) because we won't display these with
a strike-through either (though whether we should is arguable).