magit-rebase-insert-merge-sequence: Resolve abbreviated hash
At least some Git versions dump an abbreviated hash to
$GITDIR/rebase-merge/stopped-sha.  magit-rebase-insert-merge-sequence
passes in the stopped-sha value as STOP, and this abbreviated value is
compared to the full hashes of "done" items.  Luckily this faulty
check doesn't cause any issues because the magit-patch-id calls
immediately below it end up setting STOP to the full hash.
Nevertheless ensure that it is resolved to make the comparison work as
intended and to avoid unnecessary magit-patch-id calls.