Ensure all magit libraries are at least autoloaded
The optimal load/compile order is defined in "lisp/Makefile".

In lisp this is split across four places:

1. Libraries that are required by many other libraries are loaded
   in "magit-code.el".
2. Libraries that have to be loaded for "magit.el" are defined at
   the beginning of that file.
3. Other libraries that have to be loaded up-front (but may depend
   on "magit.el" being loaded) are loaded at the end of "magit.el".
4. Right after that an attempt is made to load "magit-autoload.el".
5. If that generated library cannot be found, then all remaining
   libraries are loaded up-front.  Autoloading them would be
   preferable, but since we cannot do that we have to load them
   to avoid all kinds of issues.

All libraries have to be loaded in one of these places.  Within
each place the libraries should be loaded in the same order as
in "lisp/Makefile".

This commit adds `require' forms for `magit-transient', `magit-wip',
and `magit-patch' and additionally fixes inconsistent ordering.

It is quite likely that a few libraries from (3) could be moved to
(5), but that is another issue.