magit-reset-worktree: Restrict with-temp-index scope
The temporary index is needed for the checkout-index call, but it
shouldn't be in effect for the WIP commits.  In particular, we don't
want magit-wip-commit-index to call write-tree with the temporary
index.  (magit-wip-commit-worktree, on the other hand, overrides the
index with its own magit-with-temp-index call, so its behavior isn't
influenced by magit-reset-worktree's temporary index.)

Also, refresh after the checkout-index call.  Otherwise the changes
are initially listed as staged because, at the time of the
checkout-index call, the temporary index is in effect.