Drop non-interactive uses of suffix commands
Suffix commands are only intended for non-interactive use now.

In the case of `magit-blame-echo' that means a tiny bit of
duplication.

In the case of `magit-commit-absorb' we deal with a special case.
This command has three "phases" one of them is implemented as a
non-interactive call to itself.  Because of that it would be even
more of a mistake to call this command non-interactively elsewhere
then it is for other suffix commands.