Remove magit-branch-arguments
The `magit-branch' transient has no infix arguments, but some of its
suffix commands still tried to get arguments.

"git branch" does a lot of different things related to branches.  Most
of its arguments are best thought of as sub-commands.  Other arguments
control how the output is formatted.  All of these arguments should
not be added to the transient.

The only argument that might make sense is "--force", but adding that
would be dangerous, especially if a user configured that argument to
be enabled by default.  Most commands that perform an action that
might have to be forced already do so with a prefix argument.