magit-delete-remote-branch-sentinel: Decrease false positives
When deleting a remote ref fails, we fall back to removing the local
remote-tracking ref because a common reason for the push to fail is
that the ref no longer exist on the remote.  But it may fail for other
reasons (e.g., can't connect to the remote or a non-zero exit of the
pre-receive hook), and we shouldn't remove the local tracking ref in
these cases.

One way to tighten the check would be to search the process buffer for
"unable to delete 'BRANCH': does not exist".  (We'd need to search for
each ref with a tailored message because we wouldn't want to remove
the remote-tracking refs for multiple refs when the push failed
because one ref does not exist.)  This method relies on the message
being stable and wouldn't work for some users because that message is
marked for translation.

Instead reach out to the remote to get a list of its branches, and
check the ref against this list.  Doing so adds an expense, but it
seems like an acceptable one given that the check is performed only on
failure and that it replaces cases where the user would need to
manually call 'git fetch --prune REMOTE' to clean up otherwise.

Although not strictly necessary, narrow the check to the specific
error code that we expect to see for the "no longer exists" failure.
The exit code for a failed connection is different, so this avoids
trying to connect to the remote with magit-remote-list-branches when
the push already failed to connect.

Fixes #3650.