magit-branch: Rearrange checkout suffixes
The "b" binding is the most fundamental of these commands and should
therefore not come last.  Many checkout+checkout commands correspond
to some create-only command; these pairs should be placed next to
each other.  Taken together this means that we need two rows for the
checkout commands.

This may look a bit wasteful when some suffixes are disabled and the
forge package is used:

Checkout                               Create                 Do
 b branch/revision    c new branch      n new branch           C configure...
 l local branch       S new spin-out    s new spin-off         m rename
                                                               x reset
                                                               k delete

But when all the suffixes is available then this even looks better:

Checkout                               Create                 Do
 b branch/revision    c new branch      n new branch           C configure...
 l local branch       S new spin-out    s new spin-off         m rename
 o new orphan         w new worktree    W new worktree         x reset
                      y pull-request    Y from pull-request    k delete